Cliente/requirementsClient.txt
spyne
lxml
flask
jsonschema
grpcio
grpcio-tools
fastapi
uvicorn
strawberry-graphql
pymongo
websocket-client
pika
requests
zeep
PyJWT
cryptography